+++ internal/handlers/dnd5e/v1alpha1/character/validation_converter.go
@@ -1,4 +1,4 @@
-package v1alpha1
+package character
 
 import (
 	"fmt"
